Removed unnecessary logic from CSVLogger and half tested.
CSVLogger provided a lot of functionality that wasn't being used (dates
appended to filenames, etc) as well as validating output filenames (who
cares - just write it to file whatever the name). Made it more testable.
Half finished with testing. Had to pull out the writer from the
CSVLogger so that I could mock it and test the delegation to it. Added
meanbean to pom as used to test getters/setters. Removed constructor as
there was a split between what was provided to constructor, but then
more config had to be performed via setters. Now everything is setter
based for consistency. This also avoids the need to have a massive
constructor (since I added the BufferedWriterFactory). Added
CloseableUtils to help close the plethora of writers (fun). Made
toString NPE safe.